Organize distributions module; fix make docs
