+++ pyro/distributions/__init__.py
@@ -2,12 +2,12 @@ from __future__ import absolute_import, division, print_function
 
 from pyro.distributions.delta import Delta
 from pyro.distributions.distribution import Distribution
+from pyro.distributions.iaf import InverseAutoregressiveFlow
+from pyro.distributions.multivariate_normal import MultivariateNormal
 from pyro.distributions.omt_mvn import OMTMultivariateNormal
 from pyro.distributions.rejector import Rejector
+from pyro.distributions.sparse_multivariate_normal import SparseMultivariateNormal
 from pyro.distributions.torch import *
-from pyro.distributions.torch.iaf import InverseAutoregressiveFlow
-from pyro.distributions.torch.multivariate_normal import MultivariateNormal
-from pyro.distributions.torch.sparse_multivariate_normal import SparseMultivariateNormal
 from pyro.distributions.torch_distribution import TorchDistribution
 
 # flake8: noqa
